Add files via upload
Austo’s management team wants to understand the demand of the buyers and trends in the US market. They want to build customer profiles based on the analysis to identify new purchase opportunities so that they can manipulate the business strategy and production to meet certain demand levels. Further, the analysis will be a good way for management to understand the dynamics of a new market. Suppose you are a Data Scientist working at the consulting firm that has been contracted by Austo. You are given the task to create buyer’s profiles for different types of cars with the available data as well as a set of recommendations for Austo. 

I will be performing  data analysis to generate useful insights that will help the automobile company to grow its business.